# Conflicts:
#	lpm-core/base/PagePrinter.php

+++ _db/dump.sql
@@ -132,7 +132,7 @@ CREATE TABLE `lpm_issue_mr` (
 DROP TABLE IF EXISTS `lpm_members`;
 CREATE TABLE `lpm_members` (
   `userId` bigint(10) NOT NULL,
-  `instanceType` tinyint(1) NOT NULL,
+  `instanceType` smallint(2) NOT NULL,
   `instanceId` bigint(19) NOT NULL,
   PRIMARY KEY (`userId`,`instanceType`,`instanceId`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
